Fixed wrong assertions and code
Fixed the incorrect assertion for test F to make a grade that would actually be an F. Fixed the code error in calculateNumericalGrade as essay_average was getting attributed to the wrong category (changed .exams to .essays). Fixed the indexing issue in computeAverage, (flipped "_" and "grades" and referenced it correctly.) All tests now run correctly.