Modified grade_calculator and grade_calculator_test
I added a safeguard for dividing my zero in computeAverage, as well as adding a test that makes sure it returns an F (0).
I added many tests to grade_calculator_test in order to reach 100% test coverage by testing the boundaries of the grades, testing an empty class with no grades added, tested for a ridiculously high score number (unnecessary due to already checking A), testing with multiple instantiations of assignments and exams, some as perfect duplicates just to see what could happen, and checked the GradeType function with its three options.